refactor(pptx): extract theme.rs from lib.rs (verbatim)
Theme + colour-map parsing: parse_theme_colors, CLR_MAP_LOGICALS, parse_clr_map_node, apply_clr_map, bake_clr_map, parse_clr_map_ovr, resolve_theme_typeface, and PptxSchemeResolver + its ooxml_common::color::ThemeResolver impl, now pub(crate). Shared XML helpers (child/attr) stay in lib.rs and are imported. lib.rs uses theme::*; parse_color_node/parse_color_node_tint (which construct PptxSchemeResolver) stay in lib.rs for the fill.rs extraction.
